LF-1554: drm/mxsfb: Fix possible overflow
Fix a possible overflow (at least a theoretical one, since a mode's
clock could never be near the MAX_INT in order to generate an overflow
when multiplied by 1000) in mxsfb_pipe_mode_valid function.
This overflow was discovered by Coverity scan.

Fixes: 7bfb905f8ed2 ("Add max-memory-bandwidth property for MXSFB")

+++ drivers/gpu/drm/mxsfb/mxsfb_drv.c
@@ -161,7 +161,7 @@ enum drm_mode_status mxsfb_pipe_mode_valid(struct drm_crtc *crtc,
 	else
 		bpp = pipe->plane.state->fb->format->depth;
 
-	bw = mode->clock * 1000;
+	bw = (u64)mode->clock * 1000;
 	bw = bw * mode->hdisplay * mode->vdisplay * (bpp / 8);
 	bw = div_u64(bw, mode->htotal * mode->vtotal);
